ref/Submission_ETL_Viddy/Submission_ETL_Viddy/requirements.txt
beautifulsoup4
requests
pandas
gspread
oauth2client
google-auth
pytest
coverage
python-dotenv
psycopg2